sql: move some function to conn_executor_ddl.go
Release note: None
