+++ pkg/sql/conn_executor_exec.go
@@ -42,7 +42,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/parser/statements"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
-	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgnotice"
 	"github.com/cockroachdb/cockroach/pkg/sql/physicalplan"
 	"github.com/cockroachdb/cockroach/pkg/sql/regions"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/asof"
@@ -2314,28 +2313,6 @@ var txnSchemaChangeErr = pgerror.Newf(
 	pgcode.FeatureNotSupported,
 	"to use multi-statement transactions involving a schema change under weak isolation levels, enable the autocommit_before_ddl setting",
 )
-
-// maybeUpgradeToSerializable checks if the statement is a schema change, and
-// upgrades the transaction to serializable isolation if it is. If the
-// transaction contains multiple statements, and an upgrade was attempted, an
-// error is returned.
-func (ex *connExecutor) maybeUpgradeToSerializable(ctx context.Context, stmt Statement) error {
-	p := &ex.planner
-	if tree.CanModifySchema(stmt.AST) {
-		if ex.state.mu.txn.IsoLevel().ToleratesWriteSkew() {
-			if !ex.extraTxnState.firstStmtExecuted {
-				if err := ex.state.setIsolationLevel(isolation.Serializable); err != nil {
-					return err
-				}
-				ex.extraTxnState.upgradedToSerializable = true
-				p.BufferClientNotice(ctx, pgnotice.Newf("setting transaction isolation level to SERIALIZABLE due to schema change"))
-			} else {
-				return txnSchemaChangeErr
-			}
-		}
-	}
-	return nil
-}
 
 // makeExecPlan creates an execution plan and populates planner.curPlan using
 // the cost-based optimizer. This is used to create the plan when executing a
